ADD: Only project author is able to delete his/her project in delete_project

+++ instructor_error_check.php
@@ -70,10 +70,11 @@ function modify_project($modf_pro_id, $modf_pro_title,
 }
 
 
-function delete_project($db_conn, $del_pro_id){
+function delete_project($db_conn, $fname, $lname, $csuid, $del_pro_id){
 	global $errors, $pro_array;
 	$table_name = "project";
 	$error_flag = 0;
+	$sucess_item=array();
 	// Analyze deleted items.
 	$pro_array = explode(",", $del_pro_id);	
 	$a = array();
@@ -93,21 +94,41 @@ function delete_project($db_conn, $del_pro_id){
 		$sql = "SELECT COUNT(*) AS count FROM {$table_name} WHERE pro_id=$item";
 		$results = mysqli_query($db_conn, $sql);
 		if (!$results) {
-			array_push($errors, "Failed to delete project {$item}.");
-			return -1;
+			array_push($errors, "Failed to access deleted project {$item}.");
+			$error_flag = -1;
+			continue;
 		}
 		$row=mysqli_fetch_assoc($results);
-		if ($row["count"] == 0) { // Project $item is not in system.
+		if ($row["count"] == 0) { 
 			array_push($errors, "Failure! Project {$item} is not in this system.");
 			$error_flag = -1;
+			continue;
 		}
-		else{ // Project $item is in system.
-			$sql = "DELETE FROM {$table_name} WHERE pro_id={$item}";
-			$results = mysqli_query($db_conn, $sql);
-			if (!$results) {
-				array_push($errors, "Failure! Meet problems when to delete project {$item}.");
-				$error_flag = -1;
-			}
+		// Check if current user is the author created this project.
+		// Only project's author can delete the project.
+		$sql = "SELECT fname, lname, csuid FROM {$table_name} WHERE pro_id={$item}";
+		$results = mysqli_query($db_conn, $sql);
+		if (!$results) {
+			array_push($errors, "Failed to access author of deleted project {$item}.");
+			$error_flag = -1;
+			continue;
+		}
+		$row=mysqli_fetch_assoc($results);
+		$strcmp_1 = strcmp(strtolower($row["fname"]), strtolower($fname));
+		$strcmp_2 = strcmp(strtolower($row["lname"]), strtolower($lname));
+		$strcmp_3 = strcmp(strtolower($row["csuid"]), strtolower($csuid));
+		if ($strcmp_1!=0 || $strcmp_2!=0 || $strcmp_3!=0){
+			array_push($errors, "Failure! You are not author of project {$item}.");
+			$error_flag = -1;
+			continue;
+		}
+		// Delete project $item.
+		$sql = "DELETE FROM {$table_name} WHERE pro_id={$item}";
+		$results = mysqli_query($db_conn, $sql);
+		if (!$results) {
+			array_push($errors, "Failure! Meet problems when to delete project {$item}.");
+			$error_flag = -1;
+			continue;
 		}
 	}
 	return $error_flag;	
@@ -154,7 +175,7 @@ if(isset($_POST['bn_sbmt'])){
 		case 'del_pro':
 			$pro_id = $_POST['pro_id'];
 			// Delete existing project(s).
-			$r_val = delete_project($db_conn, $pro_id);	
+			$r_val = delete_project($db_conn, $fname, $lname, $csuid, $pro_id);	
 			if ($r_val == -1){
 				mysqli_close($db_conn);
 				goto error_report;
